+++ CHANGELOG.md
@@ -2,6 +2,8 @@ CHANGELOG
 =========
 
 4.1.0 (XXXX-XX-XX)
+------------------
+
 * Add "Login with Telegram" feature (@bigfoot90)
 
 4.0.0 (2019-11-21)